Upgrade API compatibility test to Python 3.11 & undo skipped tests
This change bumps the API goldens up to their output when run with Python 3.11,
and also updates all CI to use Python 3.11 for testing the API compatibility test.
Compatibility with older Python versions is not important, because the sole
purpose of the API compatibility test is to enforce that changes to the TensorFlow
API are reviewed by an API owner. TensorFlow's compatibility itself for other Python versions is not affected.

Additionally, the code_check tests now run the API compatibility check explicitly,
while test suites now exclude the API compatibility check. This elevates the check
in importance and allows us to specifically choose the Python version to use to
run it, which is relevant right now because Python 3.11 invocations are
incompatible with the older Python versions, due to some minor changes in Python
that end up creating differences in the generated proto files.

For API reviewers: changes from Python 3.11 include (see https://docs.python.org/3/whatsnew/3.11.html)

- `add_note` is new for all exceptions
- `EnumMeta` is now `EnumType`
- `from_builtin` was removed from Signature objects

PiperOrigin-RevId: 592299026